정수형으로만 다루게 되어있어서, 실수형으로 변경

+++ src/main/java/Operator.java
@@ -5,5 +5,5 @@ public abstract class Operator {
         this.operator = operator;
     }
 
-    public abstract int getResult(int a, int b);
+    public abstract float getResult(float a, float b);
 }
